requirements.txt
pandas
numpy
camelot-py
tabula-py
pytesseract
opencv-python
pillow
pdf2image
